Move debug drawing to separate system
